Resolve the complexity-addressability trade-off on both lattices
Dense sweeps at 400 clamps per condition, every distinct field reach, both lattices. Curves are
well behaved by the criteria set in advance: split-half discrepancy across disjoint halves of the
clamps is 0.003 to 0.013 of curve span at 30x30, and jaggedness -- mean absolute second difference
over range -- is 0.045 for controlled variance, 0.067 for capacity, 0.074 for complexity, roughly
half what the sparse sweep gave.

The trade-off is real but confined to the shortest reaches, and outside it there is no trade at
all. Taking the Pareto frontier of depth-fair CX against depth-fair ADDR:

  30x30, 19 distinct reaches, 3 on the frontier
     reach 4.0    CX  39.5   ADDR 21.10   16.7 bits
     reach 15.5   CX 123.9   ADDR 12.95    2.5 bits
     reach 22.9   CX 220.0   ADDR  0.99    0.6 bits
  11x11, 9 distinct reaches, 2 on the frontier
     reach 4.0    CX 118.5   ADDR 75.47    7.3 bits
     reach 14.6   CX 153.3   ADDR 63.87    4.1 bits

Every other setting is dominated -- strictly worse on both axes at once. Past reach 23 the tissue
does not trade complexity for control, it simply loses both, so the interesting region is three
points wide at 30x30 and two at 11x11 and the sweep out to reach 414 buys nothing.

The small lattice dominates the large one wherever they can be compared. At reach 4 it holds three
times the depth-fair complexity and 3.6 times the control; at reach 15 it matches on complexity
and carries five times the control. Scaling the tissue costs addressability far more than
complexity, which is the question this investigation opened with.

The Pareto framing also disposes of the homogeneity artefact that has dogged every earlier metric.
At 11x11 capacity rises again to 4.4 bits at reach 92, on a tissue whose complexity has collapsed
to 1.41 -- the familiar trap of a near-uniform tissue being trivially steerable. It is not on the
frontier, because it is dominated. No floor or special case was needed to exclude it.

+++ generate_ensemble.py
@@ -38,7 +38,11 @@ parser.add_argument('--readout_iters', type=int, default=1,
                          'the flicker into the PCA as spurious dimensions.')
 parser.add_argument('--freq_range', type=str, default='(100.0,1000.0)')
 parser.add_argument('--amp_range', type=str, default='(-1.0,1.0)')
-parser.add_argument('--fieldScreenSize', type=int, default=None,
+parser.add_argument('--fieldScreenSize', type=float, default=None,
+                    # Float, not int: the parameter is a distance threshold, and because it
+                    # quantises onto the extracellular grid the only way to reach some
+                    # neighbourhoods is a fractional value. 11.7 grid points per cell exists
+                    # only between 2.3 and 2.9; no integer produces it.
                     help='override the field action range in the source .dat; the sweep uses this '
                          'to vary reach while holding the clamp set fixed')
 parser.add_argument('--seed', type=int, default=42)
